Fox juggernaut motions (#89)
* Addition of MovementModes

This adds the different types of movement types, easy to use and change, slowed and dashing aren't required there YET but for now the modes do work.

(this also contains parts of the dash script, this is not needed for this one, will be used in the next commit)

* Dash attempt

Feels slightly off that the lerping may not be working, apart from that, the dash itself is working
Movement also only starts when shift is pressed, I know the issue, just unsure where to set it
Just sending this as to see if you can see any issues, easier to get someone else to look at it and spot any causes for error.

* Dashing MECHANIC completion

Removed air dashing. Issue still with where to place default speed on startup so that you dont have to press shift to start moving

* Dash created / Jump created

Dash lerps as planned, if you are slightly on an angle, you will also experience gravity forced downward which makes for some cool dodging techniques which i quite liked keeping, wasn't entirely intentional, might need to change dependancy on that

Jump made, not really much to say on this, it might need lerping too in some way as the legs are getting further away from the ground so you only end up exerting less force higher up you go, but for now ill keep it as it is.

(added the game inputs to this one too, just for testing sake T is dash.)

* dont allow for motion in air

* feat: fox changes

added icon for dash
modified values
allowed wires to hit Damageable Layer

---------

Co-authored-by: Fox <teerdhareddy@gmail.com>
Co-authored-by: FollowedFox <78442792+FollowedFox@users.noreply.github.com>