feat: v1.10.0-alpha (#131)
* fix: sync of armament

* feat: run game in background

* feat: apply zoom sensitivity only on x4 and x8

* feat: added Grauwolf and Skorpion Unit Type

* fix: deactivated alignment for non owner.

This was causing a crazy issue relative to the movement of the juggernaut for clients. Its a bit difficult to explain how crazy but assume the impossible

* fix: clients jugg self spawn themselves

It was previously spawned by the host but sometimes the clients didnt finish loading the scene. Thus, it would just fall through the ground and die

* feat: increase lobby from 4 to 5 players

* feat: scene updates